fix: cargo tests after sync module reorganization with SyscallExecutorWithIpc API updates across 12 test files, MemoryManager arguments for IPC constructors, Bytes/Vec conversions, removed non-existent SIMD imports, added bytes dependency, 29 files updated

+++ kernel/src/core/data_structures/epoch_fd.rs
@@ -263,7 +263,7 @@ mod tests {
                 path: format!("/file{}", i),
                 offset: i,
             });
-            table.insert(i, handle, 0).unwrap();
+            table.insert(i as usize, handle, 0).unwrap();
         }
 
         let mut handles = vec![];
